Allow pauses between moves and steps. Fix load/save game. Fix warnings.

+++ ai_box.cpp
@@ -375,8 +375,8 @@ void AI_Box::printBox()
         }
         fprintf (stderr, "\n");
     }
-    fprintf (stderr, "    %2d %2d %2d to win, pointers %d %d\n",
+    fprintf (stderr, "    %2d %2d %2d to win, pointers %lu %lu\n",
              m_cubesToWin [Nobody], m_cubesToWin [One], m_cubesToWin [Two],
-	     m_owners, m_values);
+	     (long) m_owners, (long) m_values);
     fprintf (stderr, "\n");
 }
